+++ packages/lavas-core-vue/core/renderer.js
@@ -9,12 +9,10 @@ import {merge} from 'lodash';
 import {createBundleRenderer} from 'vue-server-renderer';
 import VueSSRClientPlugin from 'vue-server-renderer/client-plugin';
 
-import {DEFAULT_ENTRY_NAME} from './constants';
 import {distLavasPath} from './utils/path';
 import {webpackCompile, enableHotReload} from './utils/webpack';
 import templateUtil from './utils/template';
-
-import {LAVAS_DIRNAME_IN_DIST, TEMPLATE_HTML, SERVER_BUNDLE, CLIENT_MANIFEST} from './constants';
+import {DEFAULT_ENTRY_NAME, LAVAS_DIRNAME_IN_DIST, TEMPLATE_HTML, SERVER_BUNDLE, CLIENT_MANIFEST} from './constants';
 
 export default class Renderer {
     constructor(core) {
@@ -37,20 +35,23 @@ export default class Renderer {
      * @return {string} template path
      */
     getTemplateName() {
-        return this.config.router.templateFile || TEMPLATE_HTML;
+        return TEMPLATE_HTML;
     }
 
     /**
      * get template path
      *
+     * @param {?string} entryName entryName when MPA, undefined when SPA & SSR
      * @return {string} template path
      */
-    getTemplatePath() {
-        return join(this.rootDir, `core/${this.getTemplateName()}`);
+    getTemplatePath(entryName) {
+        return entryName
+            ? join(this.rootDir, `${entryName}/${TEMPLATE_HTML}`)
+            : join(this.rootDir, `core/${TEMPLATE_HTML}`);
     }
 
     /**
-     * return ssr template
+     * return SSR template content
      *
      * @param {string} base base url
      * @return {string} templateContent
@@ -169,6 +170,11 @@ export default class Renderer {
         }
     }
 
+    /**
+     * only called in SSR mode
+     * @param {object} clientConfig client webpack config
+     * @param {object} serverConfig server webpack config
+     */
     async build(clientConfig, serverConfig) {
         this.clientConfig = clientConfig;
         this.serverConfig = serverConfig;